refactor: remove redundancy and add connectivity
