+++ lib/application/personal_information/personal_information_event.dart
@@ -1,0 +1,3 @@
+part of 'personal_information_bloc.dart';
+
+class PersonalInformationEvent {}
