+++ lib/domain/core/location.dart
@@ -1,0 +1,11 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'location.freezed.dart';
+
+@freezed
+class Location with _$Location {
+  const factory Location({
+    required String code,
+    required String name,
+  }) = _Location;
+}
