+++ lib/domain/profile/profile.dart
@@ -1,5 +1,7 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
+import '../core/location.dart';
+
 part 'profile.freezed.dart';
 
 @freezed
@@ -11,6 +13,7 @@ class Profile with _$Profile {
     required String firstName,
     required String lastName,
     required String avatar,
+    required Location location,
     String? bio,
   }) = _Profile;
