+++ lib/presentation/account_information/account_information_screen.dart
@@ -1,0 +1,55 @@
+import 'package:flutter/material.dart';
+
+import '../themes/constants.dart';
+import 'widgets/contact_info.dart';
+import 'widgets/personal_info.dart';
+
+class AccountInfoScreen extends StatefulWidget {
+  const AccountInfoScreen({super.key});
+
+  @override
+  State<AccountInfoScreen> createState() => _AccountInfoScreenState();
+}
+
+class _AccountInfoScreenState extends State<AccountInfoScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      resizeToAvoidBottomInset: false,
+      backgroundColor: kSecondaryColor,
+      body: SingleChildScrollView(
+        child: SafeArea(
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 20),
+                child: const Text(
+                  'Account information',
+                  style: TextStyle(
+                    fontSize: 28,
+                    fontWeight: FontWeight.w700,
+                    color: kPrimaryColor400,
+                  ),
+                ),
+              ),
+              Padding(
+                padding: const EdgeInsets.symmetric(
+                  horizontal: 20,
+                  vertical: 30,
+                ),
+                child: Column(
+                  children: [
+                    PersonalInformation(),
+                    const SizedBox(height: 20),
+                    ContactInformation(),
+                  ],
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
