+++ lib/presentation/demo/components_demo/components_demo_screen.dart
@@ -1,6 +1,7 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 
+import '../../../domain/core/location.dart';
 import '../../../domain/crowdaction/crowdaction.dart';
 import '../../shared_widgets/accent_chip.dart';
 import '../../shared_widgets/crowdaction_card.dart';
