+++ lib/presentation/demo/demo_screen.dart
@@ -1,12 +1,8 @@
-import 'package:auto_route/auto_route.dart';
-import '../account_info_screen/account_info_screen.dart';
+import '../account_information/account_information_screen.dart';
 import 'package:flutter/material.dart';
 
 import '../../domain/core/i_settings_repository.dart';
 import '../../infrastructure/core/injection.dart';
-import '../routes/app_routes.gr.dart';
-import '../shared_widgets/rectangle_button.dart';
-import 'components_demo/current_user_status_text.dart';
 
 class DemoPage extends StatelessWidget {
   DemoPage({super.key});
