+++ test/domain/crowdaction/crowdaction_test_fixtures.dart
@@ -1,3 +1,4 @@
+import 'package:collaction_app/domain/core/location.dart';
 import 'package:collaction_app/domain/crowdaction/crowdaction.dart';
 
 import '../../test_utilities.dart';
