+++ test/domain/profile/profile_fixture.dart
@@ -1,4 +1,6 @@
+import 'package:collaction_app/domain/core/location.dart';
 import 'package:collaction_app/domain/profile/profile.dart';
+import 'package:collaction_app/infrastructure/core/location_dto.dart';
 import 'package:collaction_app/infrastructure/profile/profile_dto.dart';
 
 const cProfileDtoJson = {
@@ -16,6 +18,10 @@ const cProfileDtoFromJson = ProfileDto(
   lastName: 'Doe',
   bio: "I am Jack!",
   avatar: 'avatar',
+  location: LocationDto(
+    code: 'NL',
+    name: 'The Netherlands',
+  ),
 );
 
 const cProfileFromJson = Profile(
@@ -24,4 +30,8 @@ const cProfileFromJson = Profile(
   lastName: 'Doe',
   bio: "I am Jack!",
   avatar: 'avatar',
+  location: Location(
+    code: 'NL',
+    name: 'The Netherlands',
+  ),
 );
