+++ test/infrastructure/crowdaction/crowdaction_dto_fixtures.dart
@@ -1,4 +1,6 @@
+import 'package:collaction_app/domain/core/location.dart';
 import 'package:collaction_app/domain/crowdaction/crowdaction.dart';
+import 'package:collaction_app/infrastructure/core/location_dto.dart';
 import 'package:collaction_app/infrastructure/crowdaction/crowdaction_dto.dart';
 
 final crowdActionDto = CrowdActionDto(
