+++ test/infrastructure/crowdaction/location_dto_fixtures.dart
@@ -1,5 +1,5 @@
-import 'package:collaction_app/domain/crowdaction/crowdaction.dart';
-import 'package:collaction_app/infrastructure/crowdaction/crowdaction_dto.dart';
+import 'package:collaction_app/domain/core/location.dart';
+import 'package:collaction_app/infrastructure/core/location_dto.dart';
 
 const tLocation = Location(
   code: 'code',
