+++ test/infrastructure/profile/profile_fixtures.dart
@@ -1,4 +1,6 @@
+import 'package:collaction_app/domain/core/location.dart';
 import 'package:collaction_app/domain/profile/profile.dart';
+import 'package:collaction_app/infrastructure/core/location_dto.dart';
 import 'package:collaction_app/infrastructure/profile/profile_dto.dart';
 
 final tProfile = Profile(
@@ -7,6 +9,10 @@ final tProfile = Profile(
   lastName: 'lastName',
   avatar: 'avatar',
   bio: 'bio',
+  location: Location(
+    code: 'NL',
+    name: 'The Netherlands',
+  ),
 );
 
 final tProfileDto = ProfileDto(
@@ -15,6 +21,10 @@ final tProfileDto = ProfileDto(
   lastName: 'lastName',
   avatar: 'avatar',
   bio: 'bio',
+  location: LocationDto(
+    code: 'NL',
+    name: 'The Netherlands',
+  ),
 );
 
 final tProfileJson = {
